isolation: sched/nohz/cputime mainline precursor commits
CQID: WIND00394943

1st collection of mainline commits that are a subset of the identified
prerequisites to being able to deploy and experiment with newest
isolation content.

These are put into features/isol -- in keeping with features/rcu.
The same restriction applies to both -- i.e. they should not be
considered stand-alone features that are applicable to any baseline,
but are ordered components of OVP.

For further details on the added content, see the isol.scc file
created with this commit.

Note that this queue is inserted _above_ the kvm patches, in
keeping with tradition of kvm patches being last.  Exported
branches in kernel-3.4.x-ovp have the option of putting isol
_after_ kvm, in order to keep those BSPs fast forward for a
little while longer, if desired.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>