rcu: integrate new content from Ingo's tip repo
CQID: WIND00393800

We take a small gamble here -- the commit IDs are listed as
"upstream" but at this moment in time, the are limited to
Ingo's tip repo.  However, the merge window opens in about
three days, and so chances of a rebase in tip (that would
change the commit IDs) is low, and so there are good odds
that these commit IDs will be in mainline exactly as-is.

These commits are the final building block for the 3.10
based experimental nohz changes.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>