arm-coretile-express-a15-a7: add cgl support
CQID: WIND00409173

1/3 [
Author: Russell King
Email: rmk+kernel@arm.linux.org.uk
Subject: ARM: decompressor: avoid speculative prefetch from non-RAM areas
Date: Thu, 10 May 2012 09:48:34 +0100

Commit 1fdc08abfa26f30fcef0ce1333e9ac6f80350f30 upstream

We setup identity MMU mappings across the entire 4GB of space, which
are permissionless because the domain is set to manager.

This unfortunately allows ARMv6 and later CPUs to speculatively
prefetch from the entire address space, which can cause undesirable
side effects if those regions contain devices.

As we setup the mappings with read/write permission, we can switch
the domain to client mode, and then use the XN bit for ARMv6 and
above to control speculative prefetch to non-RAM areas.

Reported-by: R Sricharan <r.sricharan@ti.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
]

2/3 [
Author: Will Deacon
Email: will.deacon@arm.com
Subject: ARM: 7501/1: decompressor: reset ttbcr for VMSA ARMv7 cores
Date: Fri, 24 Aug 2012 15:20:59 +0100

Commit dbece45894d3ab1baac15a96dc4e1e8e23f64a93 upstream

When enabling the MMU for ARMv7 CPUs, the decompressor does not touch
the ttbcr register, assuming that it will be zeroed (N == 0, EAE == 0).
Given that only EAE is defined as 0 for non-secure copies of the
register (and a bootloader such as kexec may leave it set to 1 anyway),
we should ensure that we reset the register ourselves before turning on
the MMU.

This patch zeroes TTBCR.EAE and TTBCR.N prior to enabling the MMU for
ARMv7 cores in the decompressor, configuring us exclusively for 32-bit
translation tables via TTBR0.

Cc: <stable@vger.kernel.org>
Acked-by: Nicolas Pitre <nico@linaro.org>
Signed-off-by: Matthew Leach <matthew.leach@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
]

3/3 [
Author: Guoqing Jiang
Email: Guoqing.Jiang@windriver.com
Subject: ARM: call disable_nonboot_cpus() from machine_shutdown()
Date: Thu, 11 Apr 2013 13:28:44 +0800

From http://lists.infradead.org/pipermail/kexec/2013-January/007635.html

On Tegra at least, this change allows kexec to work with SMP enabled.
Without this,  machine_shutdown() simply puts all CPUs into a loop. If
the code of that loop is over-written, the CPUs may hang or crash (which
I do observe in practice), or cause the kexec'd kernel not to be able to
initialize them.

This fix has the added benefit that the kexec always happens on the boot
CPU, and thus kexec mirrors the initial kernel boot as much as possible.

Signed-off-by: Stephen Warren <swarren@nvidia.com>
Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Guoqing Jiang <Guoqing.Jiang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>