ecryptfs: second ls lists empty directory
CQID: WIND00415292

1/6 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: eCryptfs: Remove extra d_delete in ecryptfs_rmdir
Date: Fri, 3 May 2013 10:30:12 +0800

commit 35ffa948b2f7bdf79e488cd496232935d095087a upstream

vfs_rmdir() already calls d_delete() on the lower dentry. That was being
duplicated in ecryptfs_rmdir() and caused a NULL pointer dereference
when NFSv3 was the lower filesystem.

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

2/6 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: eCryptfs: dput dentries returned from dget_parent
Date: Fri, 3 May 2013 10:31:12 +0800

commit dd55c89852481a0708c3fd4b48f3081f4280d9d3 upstream

Call dput on the dentries previously returned by dget_parent() in
ecryptfs_rename(). This is needed for supported eCryptfs mounts on top
of the NFSv3 client.

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

3/6 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: ecryptfs: test lower_file pointer when lower_file_mutex is locked
Date: Fri, 3 May 2013 10:32:31 +0800

commit 27992890b02d340198a3a22fc210d13684a41564 upstream

This patch prevents the lower_file pointer in the 'ecryptfs_inode_info'
structure to be checked when the mutex 'lower_file_mutex' is not locked.

Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>
Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

4/6 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: switch ecryptfs_write() to struct inode *, kill on-stack fake files
Date: Fri, 3 May 2013 10:48:18 +0800

commit 48c1e44aceca577aa35be509714bd9ec4b4c3837 upstream

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

5/6 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: eCryptfs: Remove unnecessary grow_file() function
Date: Fri, 3 May 2013 10:49:42 +0800

commit bd4f0fe8bb7c73c738e1e11bc90d6e2cf9c6e20e upstream

When creating a new eCryptfs file, the crypto metadata is written out
and then the lower file was being "grown" with 4 kB of encrypted zeroes.
I suspect that growing the encrypted file was to prevent an information
leak that the unencrypted file was empty. However, the unencrypted file
size is stored, in plaintext, in the metadata so growing the file is
unnecessary.

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

6/6 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: eCryptfs: Add reference counting to lower files
Date: Fri, 3 May 2013 12:49:07 +0800

commit 332ab16f830f59e7621ae8eb2c353dc135a316f6 upstream

For any given lower inode, eCryptfs keeps only one lower file open and
multiplexes all eCryptfs file operations through that lower file. The
lower file was considered "persistent" and stayed open from the first
lookup through the lifetime of the inode.

This patch keeps the notion of a single, per-inode lower file, but adds
reference counting around the lower file so that it is closed when not
currently in use. If the reference count is at 0 when an operation (such
as open, create, etc.) needs to use the lower file, a new lower file is
opened. Since the file is no longer persistent, all references to the
term persistent file are changed to lower file.

Locking is added around the sections of code that opens the lower file
and assign the pointer in the inode info, as well as the code the fputs
the lower file when all eCryptfs users are done with it.

This patch is needed to fix issues, when mounted on top of the NFSv3
client, where the lower file is left silly renamed until the eCryptfs
inode is destroyed.

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>