BSP: UPDATE Xilinx Zynq BSP (Add CGP support): update cgl
CQID: WIND00409171

1/1 [
Author: Zhong Hongbo
Email: hongbo.zhong@windriver.com
Subject: arm: Disable outer (L2) cache in kexec
Date: Thu, 25 Apr 2013 16:33:15 +0800

background:
In the exclusive cache configuration mode, the L1 data
cache of the Cortex-A9 processor and the L2 cache are
exclusive. At any time, a given address is cached in
either L1 data cache or in the L2 cache, but not in both.

Issue:
Before boot kdump kernel, the kexec kernel will jump
a section of the relocate code. And the section code
is in L2 cache. Since It do not be flushed to memory.
After turn off the MMU, the PC of arm will get an invalid
instruct.

Reason:
In kexec kernel, When num_online_cpus equal to 1, the
outer_disable will be called. In kexec boot kdump kernel,
the other CPU will be not disabled. so the num_online_cpus
have not return 1. The outer_disable function is skip.

Resolve:
Disabling L2 before __soft_restart cures the problem.

Signed-off-by: Zhong Hongbo <hongbo.zhong@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>