isol: start updating commit IDs from tip to mainline
CQID: WIND00394943

I made a gamble that tip wasn't going to be rebased, and that paid
off, as we don't have to change the patches, now that they are
mainline.  So here we go back and update the scc file to reflect
the names of those merges, now that 3.10-rc1 has been tagged, and
hence the describe output can use that new tag.

First of several; additional merges from tip that were not able
to be described via mainline tags also need to be remapped.

To be clear -- no source changes happen here.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>