Revert "patch reorder; v3.10-rc1~111 gets applied before v3.10-rc1~89"
This reverts commit 6b5a79be68197789e6eae86cd0b8a20dc4654244.

There is some merging in tip that causes confusing ordering.

For example, we have:

v3.10-rc1~111
  commit 0279b3c0ada1d78882f24acf94ac4595bd657a89
  Merge: 797994f f300213
  Author: Linus Torvalds <torvalds@linux-foundation.org>
  Date:   Thu May 2 14:56:31 2013 -0700

      Merge branch 'sched-urgent-for-linus' of git://git.kernel.org/pub/scm/linux/kernel/git/tip/tip

and...

v3.10-rc1~89
  commit 534c97b0950b1967bca1c753aeaed32f5db40264
  Merge: 64049d1 265f22a
  Author: Linus Torvalds <torvalds@linux-foundation.org>
  Date:   Sun May 5 13:23:27 2013 -0700

      Merge branch 'timers-nohz-for-linus' of git://git.kernel.org/pub/scm/linux/kernel/git/tip/tip

So clearly, 111 (May 2nd) merge was before 89 (May 5th) merge.

However, we also have this to deal with:

    # v3.10-rc1~111^2~3 --- v3.9-rc1-48-g55eaa7c
    patch sched-Avoid-cputime-scaling-overflow.patch

which is a commit from April 30th, that is to fix a problem caused by:

    # v3.10-rc1~89^2~26^2 --- v3.9-rc1-19-gd9a3c98
    patch sched-Lower-chances-of-cputime-scaling-overflow.patch

which is a commit from Feb 20th.

The fix for the error was fed to Linus before the error itself.
So leave the mis-ordering as it was, so the error comes 1st and
then the fix for it.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>