block/adrbd: fix 32 bit compilation
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@windriver.com
Subject: block/adrbd: fix 32 bit compilation
Date: Thu, 16 May 2013 23:56:30 -0400

The 32bit variant of adrbd recursively defines memcpy, which results
in the following error:

   drivers/block/adrbd.c: In function 'adr_make_request':
   drivers/block/adrbd.c:169:15: sorry, unimplemented: inlining failed in
   call to '__builtin_memcpy': recursive inlining
   drivers/block/adrbd.c:171:2: sorry, unimplemented: called from here
   drivers/block/adrbd.c:169:15: sorry, unimplemented: inlining failed in
   call to '__builtin_memcpy': recursive inlining
   drivers/block/adrbd.c:171:2: sorry, unimplemented: called from here

adrbd calls "memcpy_nt", which has a 64 bit and 32 bit variant. The 64
bit variant defines _memcpy_nt64 and maps it to memcpy_nt via a #define.
while the 32 bit variant defines memcpy_nt directly via its inline function,
which is mapped to "memcpy" via a #define (presumably as an old fallback
to an unoptimized/generic memcopy in the past). The routine calls memcpy
internally for the fallback, which results in a recursive inline of memcpy.

So we remove the #define and the 32 bit build is once again valid.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>