nfsd4: fix hang on fast-booting nfs servers
CQID: WIND00415556

1/1 [
Author: Ming Liu
Email: ming.liu@windriver.com
Subject: nfsd4: fix hang on fast-booting nfs servers
Date: Mon, 27 May 2013 13:17:31 +0800

commit 06497524589f2a7717da33969d541674e0a27da6 upstream.

The last_close field of a cache_detail is initialized to zero, so the
condition

detail->last_close < seconds_since_boot() - 30

may be false even for a cache that was never opened.

However, we want to immediately fail upcalls to caches that were never
opened: in the case of the auth_unix_gid cache, especially, which may
never be opened by mountd (if the --manage-gids option is not set), we
want to fail the upcall immediately.  Otherwise client requests will be
dropped unnecessarily on reboot.

Also document these conditions.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Ming Liu <ming.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>