e1000: prevent oops when adapter is being closed and reset simultaneously
CQID: WIND00416944

1/1 [
Author: yanjun.zhu
Email: yanjun.zhu@windriver.com
Subject: e1000: prevent oops when adapter is being closed and reset simultaneously
Date: Mon, 27 May 2013 11:25:16 +0800

This change is based on a similar change made to e1000e support in
commit bb9e44d0d0f4 ("e1000e: prevent oops when adapter is being closed
and reset simultaneously").  The same issue has also been observed
on the older e1000 cards.

Here, we have increased the RESET_COUNT value to 50 because there are too
many accesses to e1000 nic on stress tests to e1000 nic, it is not enough
to set RESET_COUT 25. Experimentation has shown that it is enough to set
RESET_COUNT 50.

Signed-off-by: yanjun.zhu <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>