fsldma: fix controller lockups
CQID: WIND00422984

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: fsldma: fix controller lockups
Date: Mon, 1 Jul 2013 14:45:37 +0800

commit f04cd40701deace2efb9edd7120e59366bda2118 upstream

Enabling poisoning in the dmapool API quickly showed that the DMA
controller was fetching descriptors that should not have been in use.
This has caused intermittent controller lockups during testing.

I have been unable to figure out the exact set of conditions which cause
this to happen. However, I believe it is related to the driver using the
hardware registers to track whether the controller is busy or not. The
code can incorrectly decide that the hardware is idle due to lag between
register writes and the hardware actually becoming busy.

To fix this, the driver has been reworked to explicitly track the state
of the hardware, rather than try to guess what it is doing based on the
register values.

This has passed dmatest with 10 threads per channel, 100000 iterations
per thread several times without error. Previously, this would fail
within a few seconds.

Signed-off-by: Ira W. Snyder <iws@ovro.caltech.edu>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

[Adjusted the original patch a little to meet our kernel's needs.]

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>