Add a scc/cfg to control the Avoton silicon fix option
CQID: WIND00364713

A silicon issue exists with the engineering samples (ES0 and ES1) of the Avoton/Rangeley SoC
which affects the way the GbE configuration registers are accessed when issuing downstream writes to
PCI configuration space. A linux patch resolves this issue by forcing a memory map cycle that will
properly initialize the access method and don't use cf8/cfc for the first 256 bytes of config space if
MMCFG is available. In order to minimize this patch's effection, add a kernel option to limit it only
for Avoton/Rangeley Boards!

Signed-off-by: fli <fupan.li@windriver.com>