Read FMan firmware from IRam
CQID: WIND00419791

1/1 [
Author: Wangrui Shen
Email: wangrui.shen@windriver.com
Subject: Read FMan firmware from IRam
Date: Wed, 31 Jul 2013 09:11:52 +0800

Normally, U-Boot embedded the FMan firmware into DTB, and pass it
to kernel. However, certain version of U-Boot doesn't do it,
which is not right. As a workaround, kernel read the old firmware
from IRam, and use it to update IRam.

This piece of logic exist in FSL QorIQ SDK 1.1, not in FSL QorIQ
SDK 1.3. So port it from 1.1 to 1.3.

Signed-off-by: Wangrui Shen <wangrui.shen@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>