kernel part : Mobile IPv6 setup will cause crash
CQID: WIND00425506

1/2 [
Author: Nicolas Dichtel
Email: nicolas.dichtel@6wind.com
Subject: ipv6: don't add link local route when there is no link local address
Date: Mon, 1 Oct 2012 23:19:14 +0000

commit 62b54dd91567686a1cb118f76a72d5f4764a86dd upstream

When an address is added on loopback (ip -6 a a 2002::1/128 dev lo), a route
to fe80::/64 is added in the main table:
  unreachable fe80::/64 dev lo  proto kernel  metric 256  error -101

This route does not match any prefix (no fe80:: address on lo). In fact,
addrconf_dev_config() will not add link local address because this function
filters interfaces by type. If the link local address is added manually, the
route to the link local prefix will be automatically added by
addrconf_add_linklocal().
Note also, that this route is not deleted when the address is removed.

After looking at the code, it seems that addrconf_add_lroute() is redundant with
addrconf_add_linklocal(), because this function will add the link local route
when the link local address is configured.

Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Feng Mu <Feng.Mu@windriver.com>
]

2/2 [
Author: Romain Kuntz
Email: r.kuntz@ipflavors.com
Subject: ipv6: use addrconf_get_prefix_route for prefix route lookup [v2]
Date: Wed, 9 Jan 2013 21:06:03 +0000

commit 21caa6622b36190a32b19dfa734822c2eb93e1fd upstream

Replace ip6_route_lookup() with addrconf_get_prefix_route() when
looking up for a prefix route. This ensures that the connected prefix
is looked up in the main table, and avoids the selection of other
matching routes located in different tables as well as blackhole
or prohibited entries.

In addition, this fixes an Opps introduced by commit 64c6d08e (ipv6:
del unreachable route when an addr is deleted on lo), that would occur
when a blackhole or prohibited entry is selected by ip6_route_lookup().
Such entries have a NULL rt6i_table argument, which is accessed by
__ip6_del_rt() when trying to lock rt6i_table->tb6_lock.

The function addrconf_is_prefix_route() is not used anymore and is
removed.

[v2] Minor indentation cleanup and log updates.

Signed-off-by: Romain Kuntz <r.kuntz@ipflavors.com>
Acked-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Acked-by: YOSHIFUJI Hideaki <yoshfuji@linux-ipv6.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Feng Mu <Feng.Mu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>