sched:rt: don't try to pull a running task
CQID: WIND00427173

1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: sched:rt: don't try to pull a running task
Date: Tue, 6 Aug 2013 10:57:41 +0800

If we happen to pull a running task on other cpu, a WARNING
will happen. But the very task will be pulled to this cpu
finally. This pulling is dangerous since the task's private
data could be broken. Fix it by skipping this kind of task.

Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>