powerpc/mm: revert the change for _PAGE_EXEC introduced by GRSecurity patch
CQID: WIND00429869

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc/mm: revert the change for _PAGE_EXEC introduced by GRSecurity patch
Date: Mon, 24 Jun 2013 16:33:51 +0800

In the patch (GRSecurity 2.9.1 -- 201207080925) the definition of
_PAGE_EXEC is changed to _PAGE_GUARDED for the 6xx platforms. But
this is not right. On the 6xx, the _PAGE_EXEC has nothing to do with
the _PAGE_GUARDED. And actually we don't support the execute permission
check on these platforms. So restore this change.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>