ltsi: refresh 3.4.x series to LTSI commit 5f05247ed77ab71.
A test treegen shows lots of refresh work needed in the atmel
patches.  So go grab the "latest" LTSI 3.4.x content. It was
refreshed in late May to 3.4.46 and largely hasn't changed
since.  This is the raw import of patches matching LTSI

  -----------------
  commit 5f05247ed77ab71efd9e10c8fea9bb9d14445494
  Author: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
  Date:   Tue May 21 12:46:18 2013 -0700

    two marzen SATA patches added
  -----------------

and the usual conversion of LTSI series --> Yocto scc file.
We'll probably need additional refreshes in using a .57
baseline, but start with this.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>