kvm: fix lapic fallout due to stable conflict
Moving to 3.4.60 adds this stable commit:

  ---------------------
   commit 2a6b0247eee46f424e032fb7431cc4700ad19ea5
   Author: Andrew Honig <ahonig@google.com>
   Date:   Fri Mar 29 09:35:21 2013 -0700

    KVM: Allow cross page reads and writes from cached translations.

    commit 8f964525a121f2ff2df948dac908dcc65be21b5b upstream.
  ---------------------

It adds an extra arg to kvm_gfn_to_hva_cache_init().  But there were
no callers of that in lapic.c in baseline 3.4.x -- and so the stable
cherry pick dropped the lapic chunk.  We add that context/call site
in the KVM uprev patch.  So re-add the chunk that the stable queue
dropped.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>