rt: Update preempt-rt kernel config
RALLY: US18070

Enable CONFIG_DEBUG_ATOMIC_SLEEP for -rt config to align with previous releases.
Disable SLAB since it is unsupported on -rt kernel. Although, it's disabled
in Kconfig, it'd better to make it disabled here as well in order to avoid
confusion.

[ mergers note: the fragment for CONFIG_DEBUG_ATOMIC_SLEEP was already
present in kernel-cache, where it was absent in the meta branch for
kernel-3.10.x ]

Signed-off-by: Yang Shi <yang.shi@windriver.com>