kernel-debug: introduce a place holder for kernel-debug
CQID: WIND00434272

The feature/kernel-debug doesn't exist in kernel-3.10.x.  So when
--with-template=feature/kernel-debug is used on the configure line
linux-windriver.do_patch fails:

| WARNING: addon feature "features/kernel-debug/kernel-debug" was not found
| ERROR: required features were not found. aborting
| ERROR. Could not update standard/base

This feature will be filled in as it was with previous kernel versions
with things like ftrace, kmemcheck and lttng.  The placeholder will
allow the bb template in wr-kernel/templates/feature/kernel-debug to be
used without an error during linux-windriver.do_patch task.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>