lxc: fix build failures in bisection caused by disjoint fix.
A patch taken off a mailing list caused a build failure.  But
the fix for it did not come until many commits later, which is
a royal pain for anyone who wants to bisect.  This squishes the
build fix into the root cause commit of the build fail, so that
future OVP trees won't suffer this issue.

The commit in question also causes a lockdep splat; that issue
is tracked in WIND00435933 and remains to be resolved.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>