oom: avoid killing init if it assume the oom killed thread's mm
CQID: WIND00434378

1/2 [
Author: Ming Liu
Email: ming.liu@windriver.com
Subject: oom: avoid killing kthreads if they assume the oom killed thread's mm
Date: Wed, 18 Sep 2013 14:57:35 +0800

commit 7b0d44fa49b1dcfdcf4897f12ddd12ddeab1a9d7 upstream.

After selecting a task to kill, the oom killer iterates all processes and
kills all other threads that share the same mm_struct in different thread
groups.  It would not otherwise be helpful to kill a thread if its memory
would not be subsequently freed.

A kernel thread, however, may assume a user thread's mm by using
use_mm().  This is only temporary and should not result in sending a
SIGKILL to that kthread.

This patch ensures that only user threads and not kthreads are sent a
SIGKILL if they share the same mm_struct as the oom killed task.

Signed-off-by: David Rientjes <rientjes@google.com>
Reviewed-by: Michal Hocko <mhocko@suse.cz>
Reviewed-by: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Ming Liu <ming.liu@windriver.com>
]

2/2 [
Author: Ming Liu
Email: ming.liu@windriver.com
Subject: oom: avoid killing init if it assume the oom killed thread's mm
Date: Sun, 22 Sep 2013 10:25:18 +0800

After selecting a task to kill, the oom killer iterates all processes and
kills all other user threads that share the same mm_struct in different
thread groups.

But in some extreme cases, the selected task happens to be a vfork child
of init process sharing the same mm_struct with it, which causes kernel
panic on init getting killed. This panic is observed in a busybox shell
that busybox itself is init, with a kthread keeps consuming memories.

Signed-off-by: Ming Liu <ming.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>