kvm: fix hideous mangulation in uprev-to-3.6 patch
Somehow this giant monolith turd was claiming to create
arch/x86/kvm/trace.h -- which already exists.  Looking at
the old trees, it appears that the creation chunk was just
silently getting dropped on the floor.

Here, we have removed the original file before applying, which
allows it to apply, and then have captured the actual delta
in the commit.  So it now represents what was probably intended.

The delta added two trace points, and a trivial formatting
change, so it probably wasn't significantly impacting anything.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>