rcu: clobber subdir paths
We know the version the commit comes from without needing to use
subdirs.

Further, integration of the isol/ feature content into rcu will
be complicated if we need to subdivide the isol/ content by
version.  So just clobber the subdirs.

This is done w/o the "git mv ..." in order to separate the reviewable
content from the mundane moves.  Don't try and generate a tree based
off this commit.`

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>